+++ src/Settings.ts
@@ -2,7 +2,7 @@ import { App, Notice, PluginSettingTab, Setting, normalizePath } from 'obsidian'
 import SettingsProfilesPlugin from './main';
 import { DEFAULT_PROFILE_SETTINGS } from './interface';
 import { loadProfileData } from './util/SettingsFiles';
-import { ProfileSettingsModal } from './ProfileSettingsModal';
+import { ProfileSettingsModal } from './modals/ProfileSettingsModal';
 export class SettingsProfilesSettingTab extends PluginSettingTab {
 	plugin: SettingsProfilesPlugin;
 	profilesSettings: Setting[];
